msbic: remove workqueue and metrics, add helpers
This commit removes the old workqueue based system. The MSBIC will now
sync machine resources of a given type in a single goroutine. This helps
clear up some ordering issues when posting error/progress conditions.
This commit also breaks out some of the helpers into separate files to
clean up organiztion a bit and removes the old metric based alert error
mechanism.
