+++ pkg/controller/machine-set-boot-image/machine_set_boot_image_controller.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"reflect"
-	"strings"
 	"time"
 
 	opv1 "github.com/openshift/api/operator/v1"
@@ -16,12 +15,9 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	kruntime "k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/jsonmergepatch"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/apimachinery/pkg/util/wait"
 	coreinformersv1 "k8s.io/client-go/informers/core/v1"
 	clientset "k8s.io/client-go/kubernetes"
@@ -29,14 +25,10 @@ import (
 	corelisterv1 "k8s.io/client-go/listers/core/v1"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/tools/record"
-	"k8s.io/client-go/util/workqueue"
 	"k8s.io/klog/v2"
 	"k8s.io/kubectl/pkg/scheme"
-	"sigs.k8s.io/yaml"
 
 	machinev1beta1 "github.com/openshift/api/machine/v1beta1"
-	ctrlcommon "github.com/openshift/machine-config-operator/pkg/controller/common"
-
 	machineclientset "github.com/openshift/client-go/machine/clientset/versioned"
 	mapimachineinformers "github.com/openshift/client-go/machine/informers/externalversions/machine/v1beta1"
 	machinelisters "github.com/openshift/client-go/machine/listers/machine/v1beta1"
@@ -44,11 +36,7 @@ import (
 
 	mcopinformersv1 "github.com/openshift/client-go/operator/informers/externalversions/operator/v1"
 	mcoplistersv1 "github.com/openshift/client-go/operator/listers/operator/v1"
-
-	archtranslater "github.com/coreos/stream-metadata-go/arch"
-	"github.com/coreos/stream-metadata-go/stream"
-
-	osconfigv1 "github.com/openshift/api/config/v1"
+	ctrlcommon "github.com/openshift/machine-config-operator/pkg/controller/common"
 )
 
 // Controller defines the machine-set-boot-image controller.
@@ -68,16 +56,10 @@ type Controller struct {
 	infraListerSynced          cache.InformerSynced
 	mcopListerSynced           cache.InformerSynced
 
-	syncHandler func(ms string) error
-
-	queue workqueue.RateLimitingInterface
-
 	featureGateAccess featuregates.FeatureGateAccess
 }
 
 const (
-	// maxRetries is the number of times a machineset will be retried before dropping out of queue
-	maxRetries = 5
 	// Name of machine api namespace
 	MachineAPINamespace = "openshift-machine-api"
 
@@ -112,7 +94,6 @@ func New(
 		machineClient: machineClient,
 		mcopClient:    mcopClient,
 		eventRecorder: ctrlcommon.NamespacedEventRecorder(eventBroadcaster.NewRecorder(scheme.Scheme, corev1.EventSource{Component: "machineconfigcontroller-machinesetbootimagecontroller"})),
-		queue:         workqueue.NewNamedRateLimitingQueue(workqueue.DefaultControllerRateLimiter(), "machineconfigcontroller-machinesetbootimagecontroller"),
 	}
 
 	mapiMachineSetInformer.Informer().AddEventHandler(cache.ResourceEventHandlerFuncs{
@@ -130,8 +111,6 @@ func New(
 		UpdateFunc: ctrl.updateMachineConfiguration,
 	})
 
-	ctrl.syncHandler = ctrl.syncMachineResource
-
 	ctrl.mcoCmLister = mcoCmInfomer.Lister()
 	ctrl.mapiMachineSetLister = mapiMachineSetInformer.Lister()
 	ctrl.infraLister = infraInformer.Lister()
@@ -148,9 +127,8 @@ func New(
 }
 
 // Run executes the machine-set-boot-image controller.
-func (ctrl *Controller) Run(workers int, stopCh <-chan struct{}) {
+func (ctrl *Controller) Run(stopCh <-chan struct{}) {
 	defer utilruntime.HandleCrash()
-	defer ctrl.queue.ShutDown()
 
 	if !cache.WaitForCacheSync(stopCh, ctrl.mcoCmListerSynced, ctrl.mapiMachineSetListerSynced, ctrl.infraListerSynced, ctrl.mcopListerSynced) {
 		return
@@ -159,60 +137,7 @@ func (ctrl *Controller) Run(workers int, stopCh <-chan struct{}) {
 	klog.Info("Starting MachineConfigController-MachineSetBootImageController")
 	defer klog.Info("Shutting down MachineConfigController-MachineSetBootImageController")
 
-	for i := 0; i < workers; i++ {
-		go wait.Until(ctrl.worker, time.Second, stopCh)
-	}
-
 	<-stopCh
-}
-
-// worker runs a worker thread that just dequeues items, processes them, and marks them done.
-// It enforces that the syncHandler is never invoked concurrently with the same key.
-func (ctrl *Controller) worker() {
-	for ctrl.processNextWorkItem() {
-	}
-}
-
-func (ctrl *Controller) processNextWorkItem() bool {
-	key, quit := ctrl.queue.Get()
-	if quit {
-		return false
-	}
-	defer ctrl.queue.Done(key)
-
-	err := ctrl.syncHandler(key.(string))
-	ctrl.handleErr(err, key)
-
-	return true
-}
-
-func (ctrl *Controller) handleErr(err error, key interface{}) {
-
-	_, msName, splitErr := cache.SplitMetaNamespaceKey(key.(string))
-	if splitErr != nil {
-		klog.V(2).Infof("Exception %v handling error during machineset sync %v: %v", splitErr, key, err)
-		return
-	}
-
-	if err == nil {
-		ctrl.queue.Forget(key)
-		// Clear any errors as this machineset was successfully patched
-		ctrlcommon.MCCBootImageUpdateErr.WithLabelValues(msName).Set(0)
-		return
-	}
-
-	// Increment error on this machineset
-	ctrlcommon.MCCBootImageUpdateErr.WithLabelValues(msName).Inc()
-	if ctrl.queue.NumRequeues(key) < maxRetries {
-		klog.V(2).Infof("Error syncing machineset %v: %v", key, err)
-		ctrl.queue.AddRateLimited(key)
-		return
-	}
-
-	utilruntime.HandleError(err)
-	klog.V(2).Infof("Dropping machineset %q out of the work queue: %v", key, err)
-	ctrl.queue.Forget(key)
-	ctrl.queue.AddAfter(key, 1*time.Minute)
 }
 
 func (ctrl *Controller) addMAPIMachineSet(obj interface{}) {
@@ -224,10 +149,7 @@ func (ctrl *Controller) addMAPIMachineSet(obj interface{}) {
 	// Update/Check all machinesets instead of just this one. This prevents needing to maintain a local
 	// store of machineset conditions. As this is using a lister, it is relatively inexpensive to do
 	// this.
-	err := ctrl.enqueueMAPIMachineSets()
-	if err != nil {
-		klog.Errorf("Error enqueuing MAPI machinesets: %v", err)
-	}
+	go func() { ctrl.syncMAPIMachineSets() }()
 }
 
 func (ctrl *Controller) updateMAPIMachineSet(oldMS, newMS interface{}) {
@@ -235,8 +157,8 @@ func (ctrl *Controller) updateMAPIMachineSet(oldMS, newMS interface{}) {
 	oldMachineSet := oldMS.(*machinev1beta1.MachineSet)
 	newMachineSet := newMS.(*machinev1beta1.MachineSet)
 
-	// Only take action if the there is an actual change in the MachineSet's Spec, or a change in the labels
-	if oldMachineSet.Generation == newMachineSet.Generation && reflect.DeepEqual(oldMachineSet.Labels, newMachineSet.Labels) {
+	// Don't take action if the there is no change in the MachineSet's ProviderSpec, labels and ownerreferences
+	if reflect.DeepEqual(oldMachineSet.Spec.Template.Spec.ProviderSpec, newMachineSet.Spec.Template.Spec.ProviderSpec) && reflect.DeepEqual(oldMachineSet.GetLabels(), newMachineSet.GetLabels()) && reflect.DeepEqual(oldMachineSet.GetOwnerReferences(), newMachineSet.GetOwnerReferences()) {
 		return
 	}
 
@@ -245,10 +167,7 @@ func (ctrl *Controller) updateMAPIMachineSet(oldMS, newMS interface{}) {
 	// Update all machinesets instead of just this one. This prevents needing to maintain a local
 	// store of machineset conditions. As this is using a lister, it is relatively inexpensive to do
 	// this.
-	err := ctrl.enqueueMAPIMachineSets()
-	if err != nil {
-		klog.Errorf("Error enqueuing MAPI machinesets: %v", err)
-	}
+	go func() { ctrl.syncMAPIMachineSets() }()
 }
 
 func (ctrl *Controller) addConfigMap(obj interface{}) {
@@ -257,17 +176,14 @@ func (ctrl *Controller) addConfigMap(obj interface{}) {
 
 	// Take no action if this isn't the "golden" config map
 	if configMap.Name != ctrlcommon.BootImagesConfigMapName {
-		klog.V(4).Infof("configMap %s added, but does not match %s, skipping bootimage sync", configMap.Name, ctrlcommon.BootImagesConfigMapName)
 		return
 	}
 
 	klog.Infof("configMap %s added, reconciling enrolled machine resources", configMap.Name)
 
-	// Update all machinesets since the "golden" configmap has been updated
-	err := ctrl.enqueueMAPIMachineSets()
-	if err != nil {
-		klog.Errorf("Error enqueuing MAPI machinesets: %v", err)
-	}
+	// Update all machinesets since the "golden" configmap has been added
+	// TODO: Add go routines for CAPI resources here
+	go func() { ctrl.syncMAPIMachineSets() }()
 }
 
 func (ctrl *Controller) updateConfigMap(oldCM, newCM interface{}) {
@@ -277,7 +193,6 @@ func (ctrl *Controller) updateConfigMap(oldCM, newCM interface{}) {
 
 	// Take no action if this isn't the "golden" config map
 	if oldConfigMap.Name != ctrlcommon.BootImagesConfigMapName {
-		klog.V(4).Infof("configMap %s updated, but does not match %s, skipping bootimage sync", oldConfigMap.Name, ctrlcommon.BootImagesConfigMapName)
 		return
 	}
 
@@ -289,10 +204,8 @@ func (ctrl *Controller) updateConfigMap(oldCM, newCM interface{}) {
 	klog.Infof("configMap %s updated, reconciling enrolled machine resources", oldConfigMap.Name)
 
 	// Update all machinesets since the "golden" configmap has been updated
-	err := ctrl.enqueueMAPIMachineSets()
-	if err != nil {
-		klog.Errorf("Error enqueuing MAPI machinesets: %v", err)
-	}
+	// TODO: Add go routines for CAPI resources here
+	go func() { ctrl.syncMAPIMachineSets() }()
 }
 
 func (ctrl *Controller) addMachineConfiguration(obj interface{}) {
@@ -308,10 +221,8 @@ func (ctrl *Controller) addMachineConfiguration(obj interface{}) {
 	klog.Infof("Bootimages management configuration has been added, reconciling enrolled machine resources")
 
 	// Update/Check machinesets since the boot images configuration knob was updated
-	err := ctrl.enqueueMAPIMachineSets()
-	if err != nil {
-		klog.Errorf("Error enqueuing MAPI machinesets: %v", err)
-	}
+	// TODO: Add go routines for CAPI resources here
+	go func() { ctrl.syncMAPIMachineSets() }()
 }
 
 func (ctrl *Controller) updateMachineConfiguration(oldMC, newMC interface{}) {
@@ -333,129 +244,114 @@ func (ctrl *Controller) updateMachineConfiguration(oldMC, newMC interface{}) {
 	klog.Infof("Bootimages management configuration has been updated, reconciling enrolled machine resources")
 
 	// Update all machinesets since the boot images configuration knob was updated
-	err := ctrl.enqueueMAPIMachineSets()
-	if err != nil {
-		klog.Errorf("Error enqueuing MAPI machinesets: %v", err)
-	}
+	// TODO: Add go routines for CAPI resources here
+	go func() { ctrl.syncMAPIMachineSets() }()
 }
 
-// This function adds MAPI machinesets to the workqueue
-func (ctrl *Controller) enqueueMAPIMachineSet(ms *machinev1beta1.MachineSet) {
-	key, err := cache.DeletionHandlingMetaNamespaceKeyFunc(ms)
-	if err != nil {
-		utilruntime.HandleError(fmt.Errorf("couldn't get key for object %#v: %w", ms, err))
-		return
-	}
-
-	ctrl.queue.Add(key)
-}
-
-// syncMachineSet will attempt to enqueue every machineset
-func (ctrl *Controller) enqueueMAPIMachineSets() error {
+// syncMAPIMachineSets will attempt to enqueue every machineset
+func (ctrl *Controller) syncMAPIMachineSets() {
 
 	// Grab the global operator knobs
 	mcop, err := ctrl.mcopLister.Get(ctrlcommon.MCOOperatorKnobsObjectName)
 	if err != nil {
-		return fmt.Errorf("failed to fetch MachineConfiguration knobs list while enqueueing MAPI MachineSets %w", err)
+		if apierrors.IsNotFound(err) {
+			// This typically means the cache for the knobs lister is empty; don't error on this. If this
+			// object doesn't exist, something else has gone wrong in the operator's syncMachineConfiguration loop
+			klog.Infof("MachineConfiguration knobs was not found, so no MAPI machinesets will be enqueued.")
+		} else {
+			klog.Errorf("failed to fetch MachineConfiguration knobs while enqueueing MAPI MachineSets %v", err)
+		}
+		return
 	}
 
 	// If no machine managers exist; exit the enqueue process.
 	if len(mcop.Spec.ManagedBootImages.MachineManagers) == 0 {
 		klog.Infof("No MAPI machineset manager was found, so no MAPI machinesets will be enqueued.")
-		return nil
+		return
 	}
 
 	machineManagerFound, machineResourceSelector, err := getMachineResourceSelectorFromMachineManagers(mcop.Spec.ManagedBootImages.MachineManagers, opv1.MachineAPI, opv1.MachineSets)
 	if err != nil {
-		return fmt.Errorf("failed to create a machineset selector while enqueueing MAPI machineset %w", err)
+		klog.Errorf("failed to create a machineset selector while enqueueing MAPI machineset %v", err)
+		return
 	}
 	if !machineManagerFound {
 		klog.Infof("No MAPI machineset manager was found, so no MAPI machinesets will be enqueued.")
-		return nil
 	}
 
-	machineSets, err := ctrl.mapiMachineSetLister.List(machineResourceSelector)
+	mapiMachineSets, err := ctrl.mapiMachineSetLister.List(machineResourceSelector)
 	if err != nil {
-		return fmt.Errorf("failed to fetch MachineSet list while enqueueing MAPI MachineSets %w", err)
+		klog.Errorf("failed to fetch MachineSet list while enqueueing MAPI MachineSets %v", err)
+		return
 	}
 
-	klog.Infof("Reconciling %d MAPI MachineSet(s)", len(machineSets))
-	for _, machineSet := range machineSets {
-		ctrl.enqueueMAPIMachineSet(machineSet)
+	if len(mapiMachineSets) == 0 {
+		klog.Infof("No MAPI machinesets were enrolled, so no MAPI machinesets will be enqueued.")
+		return
 	}
 
-	return nil
-}
+	klog.Infof("Reconciling %d MAPI MachineSet(s)", len(mapiMachineSets))
 
-// This function checks if an array of machineManagers contains the target apigroup/resource and returns
-// a bool(success/fail), a label selector to filter the target resource and an error, if any.
-func getMachineResourceSelectorFromMachineManagers(machineManagers []opv1.MachineManager, apiGroup opv1.MachineManagerMachineSetsAPIGroupType, resource opv1.MachineManagerMachineSetsResourceType) (bool, labels.Selector, error) {
-
-	for _, machineManager := range machineManagers {
-		if machineManager.APIGroup == apiGroup && machineManager.Resource == resource {
-			if machineManager.Selection.Mode == opv1.Partial {
-				selector, err := metav1.LabelSelectorAsSelector(machineManager.Selection.Partial.MachineResourceSelector)
-				return true, selector, err
-			} else if machineManager.Selection.Mode == opv1.All {
-				return true, labels.Everything(), nil
-			}
+	for _, machineSet := range mapiMachineSets {
+		err := ctrl.syncMAPIMachineSet(machineSet)
+		if err != nil {
+			klog.Errorf("Error syncing MAPI MachineSet %v", err)
 		}
 	}
-	return false, nil, nil
 }
 
-// syncMachineResource will attempt to reconcile the machineresource with the given key.
-// This function is not meant to be invoked concurrently with the same key.
-func (ctrl *Controller) syncMachineResource(key string) error {
+// syncMAPIMachineSet will attempt to reconcile the provided machineset
+func (ctrl *Controller) syncMAPIMachineSet(machineSet *machinev1beta1.MachineSet) error {
 
 	// TODO: lower this level of all info logs in this function
 	startTime := time.Now()
-	klog.V(4).Infof("Started syncing machineset %q (%v)", key, startTime)
+	klog.V(4).Infof("Started syncing MAPI machineset %q (%v)", machineSet.Name, startTime)
 	defer func() {
-		klog.V(4).Infof("Finished syncing machineset %q (%v)", key, time.Since(startTime))
+		klog.V(4).Infof("Finished syncing MAPI machineset %q (%v)", machineSet.Name, time.Since(startTime))
 	}()
 
-	// Fetch the bootimage configmap
-	configMap, err := ctrl.mcoCmLister.ConfigMaps(ctrlcommon.MCONamespace).Get(ctrlcommon.BootImagesConfigMapName)
-	if configMap == nil || err != nil {
-		return fmt.Errorf("failed to fetch coreos-bootimages config map during machineset sync: %w", err)
-	}
-
-	// Take no action if the MCO hash version stored in the configmap does not match the current controller
+	// Wait until the MCO hash version stored in the configmap matches the current MCO
 	// version. This is done by the operator when a master node successfully updates to a new image. This is
 	// to prevent machinesets from being updated before the operator itself has updated.
+	// Could return an error(and cause degrade) immediately here, but seems excessive. Waiting with a timeout
+	// is a bit more graceful.
+	var configMap *corev1.ConfigMap
+	var err error
+	if err := wait.PollUntilContextTimeout(context.TODO(), 1*time.Minute, 15*time.Minute, true, func(ctx context.Context) (bool, error) {
+		// Fetch the bootimage configmap
+		configMap, err = ctrl.mcoCmLister.ConfigMaps(ctrlcommon.MCONamespace).Get(ctrlcommon.BootImagesConfigMapName)
+		if configMap == nil || err != nil {
+			err = fmt.Errorf("failed to fetch coreos-bootimages config map during machineset sync: %w", err)
+			return false, nil
+		}
+		versionHashFromCM, versionHashFound := configMap.Data[ctrlcommon.MCOVersionHashKey]
+		if !versionHashFound {
+			err = fmt.Errorf("failed to find mco version hash in %s configmap, sync will exit to wait for the MCO upgrade to complete", ctrlcommon.BootImagesConfigMapName)
+			return false, nil
+		}
+		if versionHashFromCM != operatorversion.Hash {
+			err = fmt.Errorf("mismatch between MCO hash version stored in configmap and current MCO version; sync will exit to wait for the MCO upgrade to complete")
+			return false, nil
+		}
+		releaseVersionFromCM, releaseVersionFound := configMap.Data[ctrlcommon.MCOReleaseImageVersionKey]
+		if !releaseVersionFound {
+			err = fmt.Errorf("failed to find mco release version in %s configmap, sync will exit to wait for the MCO upgrade to complete", ctrlcommon.BootImagesConfigMapName)
+			return false, nil
+		}
+		if releaseVersionFromCM != operatorversion.ReleaseVersion {
+			err = fmt.Errorf("mismatch between MCO release version stored in configmap and current MCO release version; sync will exit to wait for the MCO upgrade to complete")
+			return false, nil
+		}
+		return true, nil
 
-	versionHashFromCM, versionHashFound := configMap.Data[ctrlcommon.MCOVersionHashKey]
-	if !versionHashFound {
-		return fmt.Errorf("failed to find mco version hash in %s configmap, sync will exit to wait for the MCO upgrade to complete", ctrlcommon.BootImagesConfigMapName)
-	}
-	if versionHashFromCM != operatorversion.Hash {
-		return fmt.Errorf("mismatch between MCO hash version stored in configmap and current MCO version; sync will exit to wait for the MCO upgrade to complete")
-	}
-	releaseVersionFromCM, releaseVersionFound := configMap.Data[ctrlcommon.MCOReleaseImageVersionKey]
-	if !releaseVersionFound {
-		return fmt.Errorf("failed to find mco release version in %s configmap, sync will exit to wait for the MCO upgrade to complete", ctrlcommon.BootImagesConfigMapName)
-	}
-	if releaseVersionFromCM != operatorversion.ReleaseVersion {
-		return fmt.Errorf("mismatch between MCO release version stored in configmap and current MCO release version; sync will exit to wait for the MCO upgrade to complete")
+	}); err != nil {
+		klog.Errorf("Timed out waiting for coreos-bootimages config map: %v", err)
+		return fmt.Errorf("timed out waiting for coreos-bootimages config map: %v", err)
 	}
 
 	// TODO: Also check against the release version stored in the configmap under releaseVersion. This is currently broken as the version
 	// stored is "0.0.1-snapshot" and does not reflect the correct value. Tracked in this bug https://issues.redhat.com/browse/OCPBUGS-19824
 	// The current hash and version check should be enough to skate by for now, but fixing this would be additional safety - djoshy
-
-	// When adding support for multiple kinds of machine resources, the "key" format can be updated to enclose the kind of machine resource. Once decoded,
-	// an appropriate processing function can be called. For now only MAPI machinesets are supported, so proceed assuming that.
-	namespace, name, err := cache.SplitMetaNamespaceKey(key)
-	if err != nil {
-		return err
-	}
-
-	// Fetch the machineset
-	machineSet, err := ctrl.mapiMachineSetLister.MachineSets(namespace).Get(name)
-	if machineSet == nil || apierrors.IsNotFound(err) {
-		return fmt.Errorf("failed to fetch machineset object during machineset sync: %w", err)
-	}
 
 	// If the machineset has an owner reference, exit and report error. This means
 	// that the machineset may be managed by another workflow and should not be reconciled.
@@ -464,7 +360,7 @@ func (ctrl *Controller) syncMachineResource(key string) error {
 	}
 
 	// Fetch the architecture type of this machineset
-	arch, err := ctrl.getArchFromMachineSet(machineSet)
+	arch, err := getArchFromMachineSet(machineSet)
 	if err != nil {
 		return fmt.Errorf("failed to fetch arch during machineset sync: %w", err)
 	}
@@ -490,27 +386,6 @@ func (ctrl *Controller) syncMachineResource(key string) error {
 	return nil
 }
 
-// Returns architecture type for a given machineset
-func (ctrl *Controller) getArchFromMachineSet(machineset *machinev1beta1.MachineSet) (arch string, err error) {
-
-	// Valid set of machineset/node architectures
-	validArchSet := sets.New[string]("arm64", "s390x", "amd64", "ppc64le")
-	// Check if the annotation enclosing arch label is present on this machineset
-	archLabel, archLabelMatch := machineset.Annotations[MachineSetArchAnnotationKey]
-	if archLabelMatch {
-		// Grab arch value from the annotation and check if it is valid
-		_, archLabelValue, archLabelValueFound := strings.Cut(archLabel, ArchLabelKey)
-		if archLabelValueFound && validArchSet.Has(archLabelValue) {
-			return archtranslater.RpmArch(archLabelValue), nil
-		}
-		return "", fmt.Errorf("invalid architecture value found in annotation: %s ", archLabel)
-	}
-	// If no arch annotation was found on the machineset, default to the control plane arch.
-	// return the architecture of the node running this pod, which will always be a control plane node.
-	klog.Infof("Defaulting to control plane architecture")
-	return archtranslater.CurrentRpmArch(), nil
-}
-
 // This function patches the machineset object using the machineClient
 // Returns an error if marshsalling or patching fails.
 func (ctrl *Controller) patchMachineSet(oldMachineSet, newMachineSet *machinev1beta1.MachineSet) error {
@@ -532,222 +407,4 @@ func (ctrl *Controller) patchMachineSet(oldMachineSet, newMachineSet *machinev1b
 	}
 	klog.Infof("Successfully patched machineset %s", oldMachineSet.Name)
 	return nil
-}
-
-// This function calls the appropriate reconcile function based on the infra type
-// On success, it will return a bool indicating if a patch is required, and an updated
-// machineset object if any. It will return an error if any of the above steps fail.
-func checkMachineSet(infra *osconfigv1.Infrastructure, machineSet *machinev1beta1.MachineSet, configMap *corev1.ConfigMap, arch string) (bool, *machinev1beta1.MachineSet, error) {
-	switch infra.Status.PlatformStatus.Type {
-	case osconfigv1.AWSPlatformType:
-		return reconcileAWS(machineSet, configMap, arch)
-	case osconfigv1.AzurePlatformType:
-		return reconcileAzure(machineSet, configMap, arch)
-	case osconfigv1.BareMetalPlatformType:
-		return reconcileBareMetal(machineSet, configMap, arch)
-	case osconfigv1.AlibabaCloudPlatformType:
-		return reconcileAliBaba(machineSet, configMap, arch)
-	case osconfigv1.OpenStackPlatformType:
-		return reconcileOpenStack(machineSet, configMap, arch)
-	case osconfigv1.EquinixMetalPlatformType:
-		return reconcileEquinixMetal(machineSet, configMap, arch)
-	case osconfigv1.GCPPlatformType:
-		return reconcileGCP(machineSet, configMap, arch)
-	case osconfigv1.KubevirtPlatformType:
-		return reconcileKubevirt(machineSet, configMap, arch)
-	case osconfigv1.IBMCloudPlatformType:
-		return reconcileIBMCCloud(machineSet, configMap, arch)
-	case osconfigv1.LibvirtPlatformType:
-		return reconcileLibvirt(machineSet, configMap, arch)
-	case osconfigv1.VSpherePlatformType:
-		return reconcileVSphere(machineSet, configMap, arch)
-	case osconfigv1.NutanixPlatformType:
-		return reconcileNutanix(machineSet, configMap, arch)
-	case osconfigv1.OvirtPlatformType:
-		return reconcileOvirt(machineSet, configMap, arch)
-	case osconfigv1.ExternalPlatformType:
-		return reconcileExternal(machineSet, configMap, arch)
-	case osconfigv1.PowerVSPlatformType:
-		return reconcilePowerVS(machineSet, configMap, arch)
-	case osconfigv1.NonePlatformType:
-		return reconcileNone(machineSet, configMap, arch)
-	default:
-		return unmarshalToFindPlatform(machineSet, configMap, arch)
-	}
-}
-
-// This function unmarshals the machineset's provider spec into
-// a ProviderSpec object. Returns an error if providerSpec field is nil,
-// or the unmarshal fails
-func unmarshalProviderSpec(ms *machinev1beta1.MachineSet, providerSpec interface{}) error {
-	if ms.Spec.Template.Spec.ProviderSpec.Value == nil {
-		return fmt.Errorf("providerSpec field was empty")
-	}
-	if err := yaml.Unmarshal(ms.Spec.Template.Spec.ProviderSpec.Value.Raw, &providerSpec); err != nil {
-		return fmt.Errorf("unmarshal into providerSpec failedL %w", err)
-	}
-	return nil
-}
-
-// This function marshals the ProviderSpec object into a MachineSet object.
-// Returns an error if ProviderSpec or MachineSet is nil, or if the marshal fails
-func marshalProviderSpec(ms *machinev1beta1.MachineSet, providerSpec interface{}) error {
-	if ms == nil {
-		return fmt.Errorf("MachineSet object was nil")
-	}
-	if providerSpec == nil {
-		return fmt.Errorf("ProviderSpec object was nil")
-	}
-	rawBytes, err := json.Marshal(providerSpec)
-	if err != nil {
-		return fmt.Errorf("marshal into machineset failed: %w", err)
-	}
-	ms.Spec.Template.Spec.ProviderSpec.Value = &kruntime.RawExtension{Raw: rawBytes}
-	return nil
-}
-
-// This function unmarshals the golden stream configmap into a coreos
-// stream object. Returns an error if the unmarshal fails.
-func unmarshalStreamDataConfigMap(cm *corev1.ConfigMap, st interface{}) error {
-	if err := json.Unmarshal([]byte(cm.Data[StreamConfigMapKey]), &st); err != nil {
-		return fmt.Errorf("failed to parse CoreOS stream metadata: %w", err)
-	}
-	return nil
-}
-
-// GCP reconciliation function. Key points:
-// -GCP images aren't region specific
-// -GCPMachineProviderSpec.Disk(s) stores actual bootimage URL
-// -identical for x86_64/amd64 and aarch64/arm64
-func reconcileGCP(machineSet *machinev1beta1.MachineSet, configMap *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Reconciling MAPI machineset %s on GCP, with arch %s", machineSet.Name, arch)
-
-	// First, unmarshal the GCP providerSpec
-	providerSpec := new(machinev1beta1.GCPMachineProviderSpec)
-	if err := unmarshalProviderSpec(machineSet, providerSpec); err != nil {
-		return false, nil, err
-	}
-
-	// Next, unmarshal the configmap into a stream object
-	streamData := new(stream.Stream)
-	if err := unmarshalStreamDataConfigMap(configMap, streamData); err != nil {
-		return false, nil, err
-	}
-
-	// Construct the new target bootimage from the configmap
-	// This formatting is based on how the installer constructs
-	// the boot image during cluster bootstrap
-	newBootImage := fmt.Sprintf("projects/%s/global/images/%s", streamData.Architectures[arch].Images.Gcp.Project, streamData.Architectures[arch].Images.Gcp.Name)
-
-	// Grab what the current bootimage is, compare to the newBootImage
-	// There is typically only one element in this Disk array, assume multiple to be safe
-	patchRequired = false
-	newProviderSpec := providerSpec.DeepCopy()
-	for idx, disk := range newProviderSpec.Disks {
-		if newBootImage != disk.Image {
-			klog.Infof("New target boot image: %s", newBootImage)
-			klog.Infof("Current image: %s", disk.Image)
-			patchRequired = true
-			newProviderSpec.Disks[idx].Image = newBootImage
-		}
-	}
-
-	// For now, hardcode to the managed worker secret, until Custom Pool Booting is implemented. When that happens, this will have to
-	// respect the pool this machineset is targeted for.
-	if newProviderSpec.UserDataSecret.Name != ManagedWorkerSecretName {
-		newProviderSpec.UserDataSecret.Name = ManagedWorkerSecretName
-		patchRequired = true
-	}
-
-	// If patch is required, marshal the new providerspec into the machineset
-	if patchRequired {
-		newMachineSet = machineSet.DeepCopy()
-		if err := marshalProviderSpec(newMachineSet, newProviderSpec); err != nil {
-			return false, nil, err
-		}
-	}
-	return patchRequired, newMachineSet, nil
-}
-
-func reconcileAWS(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type AWS with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileAzure(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type Azure with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileBareMetal(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type BareMetal with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileAliBaba(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type AliBaba with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileOpenStack(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type OpenStack with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileEquinixMetal(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type EquinixMetal with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileKubevirt(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type Kubevirt with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileIBMCCloud(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type IBMCCloud with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileLibvirt(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type Libvirt with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileVSphere(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type VSphere with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileNutanix(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type Nutanix with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileOvirt(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type Ovirt with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcilePowerVS(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type PowerVS with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileExternal(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type External with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-func reconcileNone(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unsupported platform type None with %s arch", machineSet.Name, arch)
-	return false, nil, nil
-}
-
-// TODO - unmarshal the providerspec into each ProviderSpec type until it succeeds,
-// and then call the appropriate reconcile function. This is needed for multi platform
-// support
-func unmarshalToFindPlatform(machineSet *machinev1beta1.MachineSet, _ *corev1.ConfigMap, arch string) (patchRequired bool, newMachineSet *machinev1beta1.MachineSet, err error) {
-	klog.Infof("Skipping machineset %s, unknown platform type with %s arch", machineSet.Name, arch)
-	return false, nil, nil
 }
